+++ include/csc/context/types.hpp
@@ -45,33 +45,12 @@ struct Scope;
 using Value = std::any;
 using VariablePtr = std::shared_ptr<Variable>;
 using VariableWkr = std::weak_ptr<Variable>;
-using ScopePtr = std::shared_ptr<Scope>;
-using ScopeWkr = std::weak_ptr<Scope>;
-using VariableKeeper = std::map<String, VariablePtr>;
-using ScopeKeeper = std::map<String, ScopePtr>;
 using ValueKeeper = std::vector<Value>;
 
 struct Variable{
     String name;
     ValueKeeper values;
     ValueType type;
-};
-                                                                                                                                                                                                                                                                                                 
-struct ScopePos{
-    ScopeWkr scope;
-};
-
-struct ScopeMetaData{
-    UID id;
-    String name;
-    ScopeWkr ptr;
-};
-
-struct Scope{
-    ScopeMetaData meta;
-    ScopeWkr          parent;
-    VariableKeeper  variables;
-    ScopeKeeper     scopes;
 };
 
 struct ValueUnit{