+++ src/test/test_editor.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include <fstream>
-#include "csc/context/ctximpl/MapContext.hpp"
+#include "csc/context/Context.hpp"
 #include "csc/core/CscHandler.hpp"
 using namespace csc;
 
@@ -8,7 +8,7 @@ int main(void){
     std::ifstream ifs("myconfig.csc");
     String str((std::istreambuf_iterator<char>(ifs)), std::istreambuf_iterator<char>());
 
-    CscHandler<MapContext> handler;
+    CscHandler<ScopeType::MapScope> handler;
     CscWriter writer = handler.writer();
     CscReader reader = handler.reader();
     writer.autoEnterOn().makeScope("custom")