+++ src/test/test_iterator.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include <fstream>
-#include "csc/context/ctximpl/MapContext.hpp"
+#include "csc/context/Context.hpp"
 #include "csc/core/CscStrSeeker.hpp"
 #include "csc/syntax/CmdDriver.hpp"
 using namespace csc;
@@ -9,7 +9,7 @@ int main(void){
     std::ifstream ifs("myconfig.csc");
     String str((std::istreambuf_iterator<char>(ifs)), std::istreambuf_iterator<char>());
 
-    MapContext context;
+    Context context(ScopeType::MapScope);
     CmdDriver drv;
     CscStrSeeker seeker;
     drv.drive(str, context);